Adding more Tests, surefire and failsafe plugin

+++ Java/api/src/main/java/com/fiera/api/dtos/UserDTO.java
@@ -13,7 +13,7 @@ public class UserDTO {
 
     @Id 
     @GeneratedValue
-    private Integer UserId;
+    private Long UserId;
     @NotNull(message = "Doc Number is required")
     private String DocNumber;
     @NotNull(message = "First Name is required")
